--- client/src/components/Buttons/ButtonContainer.js
@@ -1,8 +1,0 @@
-import styled from '@emotion/styled';
-
-export const ButtonContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 240px;
-  margin-bottom: 10px;
-`;
