+++ package.json
@@ -36,6 +36,8 @@
   },
   "devDependencies": {
     "concurrently": "^5.0.2",
+    "eslint": "^6.8.0",
+    "eslint-plugin-react": "^7.17.0",
     "nodemon": "^2.0.2",
     "prettier": "^1.19.1"
   }
